requirements.txt
chromadb==0.4.15
google-genai==1.25.0
pandas==2.1.2
timeout-decorator==0.5.0
friendlywords==1.1.2
requests>=2.25.0
python-dotenv>=0.19.0
setuptools>=40.0.0
